Add a small number of unit tests
There appears to be a bug whereby if z is so far out of the
Mandelbrot set that it escapes in one iteration (which corresponds to
n==0 right now), the pixel is drawn with blue component==0xFF). This
only happens in a Release build.

Not sure why yet but the test can reproduce this (iff the test is
run with a Release configuration).