Use an array rather than a vector as the pixel buffer
Previously, it looked like accessing an element out of bounds of the
vector and caused Visual Studio and CLang to blow up when running in debug
mode.

FYI in clang, -fsanitize=address turns on the check.

Visual Studio appears to turn on an equivalent check when running in
debug mode.

gcc: apparent use -fsanitize=undefined (haven't tried it yet).